fix metadata export issues #3119
