+++ src/lua/luastorage.c
@@ -96,8 +96,8 @@ static int store_wrapper(struct dt_imageio_module_storage_t *self, struct dt_ima
 
   gchar *complete_name = g_build_filename(tmpdir, filename, (char *)NULL);
 
-  if(dt_imageio_export(imgid, complete_name, format, fdata, high_quality, upscale, FALSE, icc_type, icc_filename,
-                       icc_intent, self, self_data, num, total, NULL) != 0)
+  if(dt_imageio_export(imgid, complete_name, format, fdata, high_quality, upscale, TRUE, icc_type, icc_filename,
+                       icc_intent, self, self_data, num, total, metadata) != 0)
   {
     fprintf(stderr, "[%s] could not export to file: `%s'!\n", self->name(self), complete_name);
     g_free(complete_name);
