+++ src/views/darkroom.c
@@ -970,7 +970,7 @@ static gboolean export_key_accel_callback(GtkAccelGroup *accel_group, GObject *a
   dt_colorspaces_color_profile_type_t icc_type = dt_conf_get_int("plugins/lighttable/export/icctype");
   gchar *icc_filename = dt_conf_get_string("plugins/lighttable/export/iccprofile");
   dt_iop_color_intent_t icc_intent = dt_conf_get_int("plugins/lighttable/export/iccintent");
-  gchar *metadata_export = dt_conf_get_string("plugins/lighttable/export/metadata");
+  gchar *metadata_export = dt_lib_export_metadata_get_conf();
   // darkroom is for single images, so only export the one the user is working on
   GList *l = g_list_append(NULL, GINT_TO_POINTER(dev->image_storage.id));
   dt_control_export(l, max_width, max_height, format_index, storage_index, high_quality, upscale, style, style_append,
